The Clownfish API `Vec_Get_Size` has changed from returning `uint32_t`
to returning `size_t`.  Adapt Lucy for the effect of this change.

This branch is broken up into commits according to the kind of remedy
that was applied at each site, and the associated difficulty of proofing
the change.

This closes #36.

+++ core/Lucy/Analysis/PolyAnalyzer.c
@@ -38,7 +38,7 @@ PolyAnalyzer_init(PolyAnalyzer *self, String *language,
     PolyAnalyzerIVARS *const ivars = PolyAnalyzer_IVARS(self);
 
     if (analyzers) {
-        for (uint32_t i = 0, max = Vec_Get_Size(analyzers); i < max; i++) {
+        for (size_t i = 0, max = Vec_Get_Size(analyzers); i < max; i++) {
             CERTIFY(Vec_Fetch(analyzers, i), ANALYZER);
         }
         ivars->analyzers = (Vector*)INCREF(analyzers);
@@ -74,7 +74,7 @@ PolyAnalyzer_Transform_IMP(PolyAnalyzer *self, Inversion *inversion) {
     (void)INCREF(inversion);
 
     // Iterate through each of the analyzers in order.
-    for (uint32_t i = 0, max = Vec_Get_Size(analyzers); i < max; i++) {
+    for (size_t i = 0, max = Vec_Get_Size(analyzers); i < max; i++) {
         Analyzer *analyzer = (Analyzer*)Vec_Fetch(analyzers, i);
         Inversion *new_inversion = Analyzer_Transform(analyzer, inversion);
         DECREF(inversion);
@@ -87,7 +87,7 @@ PolyAnalyzer_Transform_IMP(PolyAnalyzer *self, Inversion *inversion) {
 Inversion*
 PolyAnalyzer_Transform_Text_IMP(PolyAnalyzer *self, String *text) {
     Vector *const   analyzers     = PolyAnalyzer_IVARS(self)->analyzers;
-    const uint32_t  num_analyzers = Vec_Get_Size(analyzers);
+    const size_t    num_analyzers = Vec_Get_Size(analyzers);
     Inversion      *retval;
 
     if (num_analyzers == 0) {
@@ -100,7 +100,7 @@ PolyAnalyzer_Transform_Text_IMP(PolyAnalyzer *self, String *text) {
     else {
         Analyzer *first_analyzer = (Analyzer*)Vec_Fetch(analyzers, 0);
         retval = Analyzer_Transform_Text(first_analyzer, text);
-        for (uint32_t i = 1; i < num_analyzers; i++) {
+        for (size_t i = 1; i < num_analyzers; i++) {
             Analyzer *analyzer = (Analyzer*)Vec_Fetch(analyzers, i);
             Inversion *new_inversion = Analyzer_Transform(analyzer, retval);
             DECREF(retval);
